+++ src/nexgen/beamlines/I19_2_nxs.py
@@ -5,13 +5,16 @@ Create a NeXus file for time-resolved collections on I19-2.
 from __future__ import annotations
 
 import logging
-from collections import namedtuple
 from datetime import datetime
+from enum import Enum
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, NamedTuple, Optional
 
 import h5py
 import numpy as np
+from pydantic import field_validator
+
+from nexgen.utils import get_iso_timestamp
 
 from .. import log
 from ..nxs_utils import (
@@ -27,12 +30,46 @@ from ..nxs_utils.scan_utils import calculate_scan_points, identify_osc_axis
 from ..nxs_write.nxmx_writer import EventNXmxFileWriter, NXmxFileWriter
 from ..tools.meta_reader import define_vds_data_type, update_axes_from_meta
 from ..tools.metafile import DectrisMetafile
-from ..utils import find_in_dict, get_iso_timestamp, get_nexus_filename
 from .beamline_utils import GeneralParams, collection_summary_log
+from .I19_2_params import I19_2Eiger, I19_2Tristan
+
+# Define a logger object
+logger = logging.getLogger("nexgen.I19-2_NeXus")
 
 
-class ExperimentTypeError(Exception):
-    pass
+# Useful axis definitions and parameters
+class GonioAxisPosition(NamedTuple):
+    """Definition of goniometer axis name, start and end position, increment.
+
+    Fields:
+        id (str): Axis name.
+        start (float): Axis start position.
+        increment (float): Axis increment value, only needed for the scan axis. Defaults to 0.0.
+        end (float, optional): Axis end position, should only be passed for Tristan (if not passed, stills \
+            will be assumed). Defaults to None.
+    """
+
+    id: str
+    start: float
+    inc: float = 0.0
+    end: float | None = None
+
+
+class DetAxisPosition(NamedTuple):
+    """Definition of detector axis name and position.
+
+    Fields:
+        id (str): Axis name.
+        start (float): Axis start position.
+    """
+
+    id: str
+    start: float = 0.0
+
+
+class DetectorName(str, Enum):
+    EIGER = "eiger"
+    TRISTAN = "tristan"
 
 
 class CollectionParams(GeneralParams):
@@ -45,42 +82,32 @@ class CollectionParams(GeneralParams):
         detector_name (str): Name of the detector in use for current experiment.
         tot_num_images (int, optional): Total number of frames in a collection.
         scan_axis (str, optional): Rotation scan axis. Must be passed for Tristan.
+        axes_pos (list[GonioAxisPosition], optional): list of (axis_name, start, end) values for the \
+            goniometer, passed from command line. Defaults to None.
+        det_pos (list[DetAxisPosition], optional): List of (axis_name, start) values for the \
+            detector, passed from command line. Defaults to None.
     """
 
-    metafile: Path | str
-    detector_name: str
-    tot_num_images: Optional[int]
-    scan_axis: Optional[str]
+    metafile: Path
+    detector_name: DetectorName
+    tot_num_images: Optional[int] = None
+    scan_axis: Optional[str] = None
+    axes_pos: Optional[list[GonioAxisPosition]] = None
+    det_pos: Optional[list[DetAxisPosition]] = None
 
-
-# Define a logger object
-logger = logging.getLogger("nexgen.I19-2_NeXus")
-
-
-# Useful axis definitions
-axes = namedtuple("axes", ("id", "start", "inc", "end"), defaults=(None, 0.0, 0.0, 0.0))
-axes.__doc__ = """Goniometer axis name, start and end position, increment."""
-axes.id.__doc__ = "Axis name."
-axes.start.__doc__ = "Axis start position. Defaults fo 0.0."
-axes.inc.__doc__ = (
-    "Axis increment value. Defaults fo 0.0. Only needed for the scan axis."
-)
-axes.end.__doc__ = (
-    "Axis end position. Defaults fo 0.0. Only really needed for Tristan collections."
-)
-det_axes = namedtuple("det_axes", ("id", "start"), defaults=(None, 0.0))
-det_axes.__doc__ = """Detector axis name and position."""
-det_axes.id.__doc__ = "Axis name."
-det_axes.start.__doc__ = "Axis position. Defaults to 0.0."
+    @field_validator("metafile", mode="before")
+    @classmethod
+    def _parse_metafile(cls, metafile: str | Path):
+        if isinstance(metafile, str):
+            return Path(metafile)
+        return metafile
 
 
 def tristan_writer(
     master_file: Path,
     TR: CollectionParams,
-    timestamps: Tuple[str, str] = (None, None),
-    axes_pos: List[axes] = None,
-    det_pos: List[det_axes] = None,
-    notes: Dict[str, Any] | None = None,
+    timestamps: tuple[str, str] = (None, None),
+    notes: dict[str, Any] | None = None,
 ):
     """
     A function to call the nexus writer for Tristan 10M detector.
@@ -88,24 +115,19 @@ def tristan_writer(
     Args:
         master_file (Path): Path to nexus file to be written.
         TR (CollectionParams): Parameters passed from the beamline.
-        timestamps (Tuple[str, str], optional): Collection start and end time. Defaults to (None, None).
-        axes_pos (List[axes], optional): List of (axis_name, start, end) values for the \
-            goniometer, passed from command line. Defaults to None.
-        det_pos (List[det_axes], optional): List of (axis_name, start) values for the \
-            detector, passed from command line. Defaults to None.
-        notes (Dict[str, Any], optional): Dictionary of (key, value) pairs where key represents the \
+        timestamps (tuple[str, str], optional): Collection start and end time. Defaults to (None, None).
+        notes (dict[str, Any], optional): Dictionary of (key, value) pairs where key represents the \
             dataset name and value its data. Defaults to None.
     """
     source = Source("I19-2")
-    from .I19_2_params import I19_2Tristan as axes_params
 
     # Define Tristan params
     tristan_params = TristanDetector("Tristan 10M", (3043, 4183))
 
     # Define Goniometer axes
-    gonio_axes = axes_params.gonio
+    gonio_axes = I19_2Tristan.gonio
     # Define Detector
-    det_axes = axes_params.det_axes
+    det_axes = I19_2Tristan.det_axes
 
     # Update axes
     # Identify scan axis
@@ -113,14 +135,14 @@ def tristan_writer(
 
     # Goniometer
     end_pos = None
-    for gax in axes_pos:
+    for gax in TR.axes_pos:
         idx = [n for n, ax in enumerate(gonio_axes) if ax.name == gax.id][0]
         gonio_axes[idx].start_pos = gax.start
         if gax.id == scan_axis and gax.start != gax.end:
             end_pos = gax.end
 
     # Detector
-    for dax in det_pos:
+    for dax in TR.det_pos:
         idx = [n for n, ax in enumerate(det_axes) if ax.name == dax.id][0]
         det_axes[idx].start_pos = dax.start
 
@@ -136,7 +158,7 @@ def tristan_writer(
         det_axes,
         TR.beam_center,
         TR.exposure_time,
-        [axes_params.fast_axis, axes_params.slow_axis],
+        [I19_2Tristan.fast_axis, I19_2Tristan.slow_axis],
     )
 
     # Define Goniometer
@@ -187,35 +209,29 @@ def tristan_writer(
 def eiger_writer(
     master_file: Path,
     TR: CollectionParams,
-    timestamps: Tuple[str, str] = (None, None),
+    timestamps: tuple[str, str] = (None, None),
     use_meta: bool = False,
     n_frames: int | None = None,
-    axes_pos: List[axes] | None = None,
-    det_pos: List[det_axes] | None = None,
     vds_offset: int = 0,
-    notes: Dict[str, Any] | None = None,
+    notes: dict[str, Any] | None = None,
 ):
     """
     A function to call the NXmx nexus file writer for Eiger 2X 4M detector.
-    If use_meta is set to False, axes_pos and det_pos become required arguments. Otherwise, \
-    axes_pos and det_pos can be None but the code requires the information contained inside \
-    the meta file to work correctly.
+    If use_meta is set to False, then the parameter fields axes_pos and det_pos become required arguments.
+    Otherwise, axes_pos and det_pos can be None but the code requires the information contained inside \
+    the meta file be correct and readable.
 
     Args:
         master_file (Path): Path to nexus file to be written.
         TR (CollectionParams): Parameters passed from the beamline.
-        timestamps (Tuple[str, str], optional): Collection start and end time. Defaults to (None, None).
+        timestamps (tuple[str, str], optional): Collection start and end time. Defaults to (None, None).
         use_meta (bool, optional): If True, metadata such as axes positions, wavelength etc. \
             will be updated using the meta.h5 file. Defaults to False.
-        num_frames (int, optional): Number of images for the nexus file. Not necessary if same as the \
-            tot_num_images from the CollectionParameters. If different, the VDS will onlu contain the \
+        num_frames (int, optional): Number of images for the nexus file. Not necessary the same as the \
+            tot_num_images from the CollectionParameters. If different, the VDS will only contain the \
             number of frames specified here. Defaults to None.
-        axes_pos (List[axes], optional): List of (axis_name, start, inc) values for the \
-            goniometer, passed from command line. Defaults to None.
-        det_pos (List[det_axes], optional): List of (axis_name, start) values for the \
-            detector, passed from command line. Defaults to None.
         vds_offset (int, optional): Start index for the vds writer. Defaults to 0.
-        notes (Dict[str, Any], optional): Dictionary of (key, value) pairs where key represents the \
+        notes (dict[str, Any], optional): Dictionary of (key, value) pairs where key represents the \
             dataset name and value its data. Defaults to None.
 
     Raises:
@@ -223,14 +239,16 @@ def eiger_writer(
         IOError: If the axes positions can't be read from the metafile (missing config or broken links).
     """
     if not use_meta:
-        if axes_pos is None or det_pos is None:
+        if TR.axes_pos is None or TR.det_pos is None:
             logger.error(
                 """
                 If not using the meta file, please pass the complete axis information for goniometer
                 and/or detector.
                 """
             )
-            raise ValueError("Missing at least one of axes_pos or det_pos.")
+            raise ValueError(
+                "No meta file selected and missing at least one of axes_pos or det_pos from parameter model."
+            )
         if n_frames is None and TR.tot_num_images is None:
             logger.error(
                 """
@@ -239,10 +257,11 @@ def eiger_writer(
                 collection, or different if the vds needs to only point to part of the dataset.
                 """
             )
-            raise ValueError("Missing number of images.")
+            raise ValueError(
+                "Neither total number of images nor number of frames have been passed to the model."
+            )
 
     source = Source("I19-2")
-    from .I19_2_params import I19_2Eiger as axes_params
 
     # Define Eiger 4M params
     eiger_params = EigerDetector(
@@ -259,9 +278,9 @@ def eiger_writer(
     beam_center = TR.beam_center
 
     # Define Goniometer axes
-    gonio_axes = axes_params.gonio
+    gonio_axes = I19_2Eiger.gonio
     # Define Detector
-    det_axes = axes_params.det_axes
+    det_axes = I19_2Eiger.det_axes
 
     # Update axes
     if use_meta:
@@ -304,14 +323,14 @@ def eiger_writer(
         vds_dtype = np.uint32
         # Update axes
         # Goniometer
-        for gax in axes_pos:
+        for gax in TR.axes_pos:
             idx = [n for n, ax in enumerate(gonio_axes) if ax.name == gax.id][0]
             gonio_axes[idx].start_pos = gax.start
             if gax.inc != 0.0:
                 gonio_axes[idx].increment = gax.inc
 
         # Detector
-        for dax in det_pos:
+        for dax in TR.det_pos:
             idx = [n for n, ax in enumerate(det_axes) if ax.name == dax.id][0]
             det_axes[idx].start_pos = dax.start
             logger.info(
@@ -355,7 +374,7 @@ def eiger_writer(
         det_axes,
         beam_center,
         TR.exposure_time,
-        [axes_params.fast_axis, axes_params.slow_axis],
+        [I19_2Eiger.fast_axis, I19_2Eiger.slow_axis],
     )
 
     # Define Goniometer
@@ -402,68 +421,35 @@ def eiger_writer(
         raise
 
 
-def nexus_writer(
-    meta_file: Path | str,
-    detector_name: str,
-    exposure_time: float,
-    scan_axis: str = "phi",
-    start_time: datetime | None = None,
-    stop_time: datetime | None = None,
-    **params,
+def serial_nexus_writer(
+    params: dict[str, Any],
+    master_file: Path,
+    timestamps: tuple[datetime, datetime] = (None, None),
+    use_meta: bool = False,
+    vds_offset: int = 0,
+    n_frames: int | None = None,
+    notes: dict[str, Any] | None = None,
 ):
-    """
-    Gather all parameters from the beamline and call the NeXus writers.
+    """Wrapper function to gather all parameters from the beamline and kick off the nexus writer for a \
+    serial experiment on I19-2.
 
     Args:
-        meta_file (Path | str): Path to _meta.h5 file.
-        detector_name (str): Detector in use.
-        exposure_time (float): Exposure time, in s.
-        scan_axis (str, optional): Name of the oscillation axis. Defaults to phi.
-        start_time (datetime, optional): Experiment start time. Defaults to None.
-        stop_time (datetime, optional): Experiment end time. Defaults to None.
-
-    Keyword Args:
-        n_imgs (int): Total number of images to be collected.
-        transmission (float): Attenuator transmission, in %.
-        wavelength (float): Wavelength of incident beam, in A.
-        beam_center (List[float, float]): Beam center position, in pixels.
-        gonio_pos (List[axes]): Name, start and end positions \
-            of the goniometer axes.
-        det_pos (List[det_axes]): Name, start and end positions \
-            of detector axes.
-        outdir (str): Directory where to save the file. Only specify if different \
-            from meta_file directory.
-        serial (bool): Specify whether it's a serial crystallography dataset.
-        det_dist (float): Distance between sample and detector, in mm.
-        use_meta (bool): For Eiger, if True use metadata from meta.h5 file. Otherwise \
-            will require all other information to be passed manually.
+        params (dict[str, Any]): Dictionary representation of CollectionParams.
+        master_file (Path): Full path to the nexus file to be written.
+        timestamps (tuple[str, str], optional): Start and end collection timestamps as datetime. \
+            Defaults to (None, None).
+        use_meta (bool, optional): Eiger option only, if True use metadata from meta.h5 file. Otherwise \
+            all parameters will need to be passed manually. Defaults to False.
+        vds_offset (int, optional): Start index for the vds writer. Defaults to 0.
+        n_frames (int | None, optional): Number of images for the nexus file. Only needed if different \
+            from the tot_num_images in the collection params. If passed, the VDS will only contain the \
+            number of frames specified here. Defaults to None.
+        notes (dict[str, Any] | None, optional): Any additional information to be written as NXnote, \
+            passed as a dictionary of (key, value) pairs where key represents the dataset name and \
+            value its data. Defaults to None.
     """
-    if find_in_dict("serial", params) and params["serial"] is True:
-        raise ExperimentTypeError(
-            "This is writer is not enabled for ssx collections."
-            "Pleas look into SSX_Eiger or SSX_Tristan for this functionality."
-        )
-
-    TR = CollectionParams(
-        metafile=Path(meta_file).expanduser().resolve(),
-        detector_name=detector_name.lower(),
-        exposure_time=exposure_time,
-        beam_center=(
-            params["beam_center"] if find_in_dict("beam_center", params) else (0, 0)
-        ),
-        wavelength=params["wavelength"] if find_in_dict("wavelength", params) else None,
-        transmission=(
-            params["transmission"] if find_in_dict("transmission", params) else None
-        ),
-        tot_num_images=params["n_imgs"] if find_in_dict("n_imgs", params) else None,
-        scan_axis=scan_axis,
-    )
-
-    # Check that the new NeXus file is to be written in the same directory
-    if find_in_dict("outdir", params) and params["outdir"]:
-        wdir = Path(params["outdir"]).expanduser().resolve()
-    else:
-        wdir = TR.metafile.parent
+    collection_params = CollectionParams(**params)
+    wdir = master_file.parent
 
     # Define a file handler
     logfile = wdir / "I19_2_nxs_writer.log"
@@ -471,49 +457,97 @@ def nexus_writer(
     log.config(logfile.as_posix())
 
     logger.info("NeXus file writer for beamline I19-2 at DLS.")
-    logger.info(f"Detector in use for this experiment: {TR.detector_name}.")
-    logger.info(f"Current collection directory: {TR.metafile.parent}")
+    logger.info(
+        f"Detector in use for this experiment: {collection_params.detector_name.value}."
+    )
+    logger.info(f"Current collection directory: {collection_params.metafile.parent}")
 
-    # Add some information to logger
-    logger.info("Creating a NeXus file for %s ..." % TR.metafile.name)
     # Get NeXus filename
-    master_file = get_nexus_filename(TR.metafile)
-    master_file = wdir / master_file.name
     logger.info("NeXus file will be saved as %s" % master_file)
 
     # Get timestamps in the correct format if they aren't already
-    start_time = start_time.strftime("%Y-%m-%dT%H:%M:%S") if start_time else None
-    stop_time = stop_time.strftime("%Y-%m-%dT%H:%M:%S") if stop_time else None
+    start_time = timestamps[0].strftime("%Y-%m-%dT%H:%M:%S") if timestamps[0] else None
+    stop_time = timestamps[1].strftime("%Y-%m-%dT%H:%M:%S") if timestamps[1] else None
     timestamps = (
         get_iso_timestamp(start_time),
         get_iso_timestamp(stop_time),
     )
 
-    if not find_in_dict("gonio_pos", params):
-        params["gonio_pos"] = None
-    if not find_in_dict("det_pos", params):
-        params["det_pos"] = None
+    match collection_params.detector_name:
+        case DetectorName.EIGER:
+            eiger_writer(
+                master_file,
+                collection_params,
+                timestamps,
+                use_meta,
+                n_frames,
+                vds_offset,
+                notes,
+            )
+        case DetectorName.TRISTAN:
+            tristan_writer(master_file, collection_params, timestamps, notes)
 
-    if "tristan" in TR.detector_name.lower():
-        if params["gonio_pos"] is None or params["det_pos"] is None:
+
+def nexus_writer(
+    params: dict[str, Any],
+    master_file: Path,
+    timestamps: tuple[datetime, datetime] = (None, None),
+    use_meta: bool = False,
+):
+    """Wrapper function to gather all parameters from the beamline and kick off the nexus writer for a \
+    standard experiment on I19-2.
+
+    Args:
+        params (dict[str, Any]): Dictionary representation of CollectionParams.
+        master_file (Path): Full path to the nexus file to be written.
+        timestamps (tuple[str, str], optional): Start and end collection timestamps as datetime. \
+            Defaults to (None, None).
+        use_meta (bool, optional): Eiger option only, if True use metadata from meta.h5 file. Otherwise \
+            all parameters will need to be passed manually. Defaults to False.
+    """
+    collection_params = CollectionParams(**params)
+    wdir = master_file.parent
+
+    # Define a file handler
+    logfile = wdir / "I19_2_nxs_writer.log"
+    # Configure logging
+    log.config(logfile.as_posix())
+
+    logger.info("NeXus file writer for beamline I19-2 at DLS.")
+    logger.info(
+        f"Detector in use for this experiment: {collection_params.detector_name}."
+    )
+    logger.info(f"Current collection directory: {collection_params.metafile.parent}")
+
+    # Add some information to logger
+    logger.info("Creating a NeXus file for %s ..." % collection_params.metafile.name)
+    logger.info("NeXus file will be saved as %s" % master_file)
+
+    # Get timestamps in the correct format if they aren't already
+    start_time = timestamps[0].strftime("%Y-%m-%dT%H:%M:%S") if timestamps[0] else None
+    stop_time = timestamps[1].strftime("%Y-%m-%dT%H:%M:%S") if timestamps[1] else None
+    timestamps = (
+        get_iso_timestamp(start_time),
+        get_iso_timestamp(stop_time),
+    )
+
+    if collection_params.detector_name is DetectorName.TRISTAN:
+        if not collection_params.axes_pos or not collection_params.det_pos:
             logger.error("Please pass the axes positions for a Tristan collection.")
             raise ValueError(
                 "Missing goniometer and/or detector axes information for tristan collection"
             )
-        if TR.scan_axis is None:
+        if collection_params.scan_axis is None:
             logger.warning(
                 "No scan axis has been specified. Phi will be set as default."
             )
+            collection_params.scan_axis = "phi"
 
-    if not find_in_dict("use_meta", params):
-        # If by any chance not passed, assume False
-        params["use_meta"] = False
-
-    if params["use_meta"] is True:
-        params["gonio_pos"] = None
-        params["det_pos"] = None
-    else:
-        if not find_in_dict("n_imgs", params) and "eiger" in TR.detector_name:
+    if not use_meta:
+        if (
+            not collection_params.tot_num_images
+            and collection_params.detector_name is DetectorName.EIGER
+        ):
             raise ValueError(
                 """
                 Missing input parameter n_imgs. \n
@@ -521,7 +555,7 @@ def nexus_writer(
                 be collected has to be passed to the writer.
                 """
             )
-        if TR.beam_center == (0, 0):
+        if collection_params.beam_center == (0, 0):
             logger.warning(
                 """
                 Beam centre was not passed to the writer.
@@ -529,28 +563,17 @@ def nexus_writer(
                 """
             )
 
-    if find_in_dict("notes", params) and isinstance(params["notes"], dict):
-        notes = params["notes"]
-
-    if "eiger" in TR.detector_name:
-        if not find_in_dict("n_imgs", params):
-            params["n_imgs"] = None
-        eiger_writer(
-            master_file,
-            TR,
-            timestamps,
-            params["use_meta"],
-            params["n_imgs"],
-            params["gonio_pos"],
-            params["det_pos"],
-            notes=notes,
-        )
-    elif "tristan" in TR.detector_name:
-        tristan_writer(
-            master_file,
-            TR,
-            timestamps,
-            params["gonio_pos"],
-            params["det_pos"],
-            notes=notes,
-        )
+    match collection_params.detector_name:
+        case DetectorName.EIGER:
+            eiger_writer(
+                master_file,
+                collection_params,
+                timestamps,
+                use_meta,
+            )
+        case DetectorName.TRISTAN:
+            tristan_writer(
+                master_file,
+                collection_params,
+                timestamps,
+            )