+++ src/nexgen/utils.py
@@ -6,10 +6,9 @@ from __future__ import annotations
 
 import logging
 import re
-from collections import namedtuple
-from datetime import datetime
+from datetime import datetime, timezone
 from pathlib import Path
-from typing import Dict, List, Tuple
+from typing import Dict, List, NamedTuple, Tuple
 
 import h5py
 import numpy as np
@@ -31,9 +30,15 @@ __all__ = [
 MAX_FRAMES_PER_DATASET = 1000
 MAX_SUFFIX_DIGITS = 6
 
+
 # Define coordinates
-Point3D = namedtuple("Point3D", ("x", "y", "z"))
-Point3D.__doc__ = """Coordinates in 3D space."""
+class Point3D(NamedTuple):
+    """Coordinates (x,y,z) in 3D space."""
+
+    x: float
+    y: float
+    z: float
+
 
 # Filename pattern: filename_######.h5 or filename_meta.h5
 # P = re.compile(r"(.*)_(?:\d+)")
@@ -210,7 +215,15 @@ def units_of_time(q: str) -> Q_:  # -> pint.Quantity:
         )
 
 
-def get_iso_timestamp(ts: str | float) -> str:
+def _validate_timestamp_string(ts: str, fmt: str) -> bool:
+    try:
+        res = bool(datetime.strptime(ts, fmt))
+    except ValueError:
+        res = False
+    return res
+
+
+def get_iso_timestamp(ts: str | float | None) -> str:
     """
     Format a timestamp string to be stores in a NeXus file according to ISO8601: 'YY-MM-DDThh:mm:ssZ'
 
@@ -223,6 +236,7 @@ def get_iso_timestamp(ts: str | float) -> str:
     """
     # Format strings for timestamps
     format_list = [
+        "%Y-%m-%dT%H:%M:%SZ",  # ISO8601 formatted string
         "%Y-%m-%dT%H:%M:%S",
         "%Y-%m-%d %H:%M:%S",
         "%a %b %d %Y %H:%M:%S",
@@ -230,15 +244,29 @@ def get_iso_timestamp(ts: str | float) -> str:
     ]
     if ts is None:
         return None
-    try:
+
+    ts_iso = None
+    if isinstance(ts, float):
         ts = float(ts)
-        ts_iso = datetime.utcfromtimestamp(ts).replace(microsecond=0).isoformat()
-    except ValueError:
+        ts_iso = (
+            datetime.fromtimestamp(ts, tz=timezone.utc)
+            .replace(microsecond=0)
+            .isoformat()
+        )
+        ts_iso = ts_iso.removesuffix("+00:00")  # Remove timezone indication
+    elif isinstance(ts, str):
         for fmt in format_list:
-            try:
+            if _validate_timestamp_string(ts, fmt) is True:
                 ts_iso = datetime.strptime(ts, fmt).isoformat()
-            except ValueError:
-                ts_iso = str(ts)
+                break
+        if not ts_iso:
+            raise ValueError(
+                f"Unknown format. Unable to validate timestamp string, please pass one of: {format_list}"
+            )
+    else:
+        raise ValueError(
+            "Please pass the timestamp either as a time.time float or a formatted string."
+        )
     if ts_iso.endswith("Z") is False:
         ts_iso += "Z"
     return ts_iso